requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
elastic-enterprise-search==7.14.0
elastic-transport==7.14.0
idna==3.2
PyJWT==2.1.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
python-dotenv~=0.19.0